refactor: make each model a single package to standardize the whole library for easier management;
