+++ pypots/__init__.py
@@ -24,13 +24,19 @@ PyPOTS package.
 # 'X.Y.dev0' is the canonical version of 'X.Y.dev'
 __version__ = "0.0.11"
 
+from pypots import classification
+from pypots import clustering
+from pypots import data
+from pypots import forecasting
+from pypots import imputation
+from pypots import utils
 
 __all__ = [
-    "data",
-    "imputation",
     "classification",
     "clustering",
+    "data",
     "forecasting",
+    "imputation",
     "utils",
     "__version__",
 ]
