+++ pypots/classification/brits/model.py
@@ -13,8 +13,11 @@ import torch.nn.functional as F
 from torch.utils.data import DataLoader
 
 from pypots.classification.base import BaseNNClassifier
-from pypots.data import DatasetForBRITS
-from pypots.imputation.brits import RITS as imputation_RITS, _BRITS as imputation_BRITS
+from pypots.classification.brits.dataset import DatasetForBRITS
+from pypots.imputation.brits.model import (
+    RITS as imputation_RITS,
+    _BRITS as imputation_BRITS,
+)
 
 
 class RITS(imputation_RITS):
