+++ pypots/classification/grud/__init__.py
@@ -1,0 +1,12 @@
+"""
+
+"""
+
+# Created by Wenjie Du <wenjay.du@gmail.com>
+# License: GLP-v3
+
+from pypots.classification.grud.model import GRUD
+
+__all__ = [
+    "GRUD",
+]
