+++ pypots/classification/raindrop/__init__.py
@@ -1,0 +1,12 @@
+"""
+
+"""
+
+# Created by Wenjie Du <wenjay.du@gmail.com>
+# License: GLP-v3
+
+from pypots.classification.raindrop.model import Raindrop
+
+__all__ = [
+    "Raindrop",
+]
