+++ pypots/data/__init__.py
@@ -6,9 +6,6 @@ Expose all usable data manipulation classes and functions.
 # License: GPL-v3
 
 from pypots.data.base import BaseDataset
-from pypots.data.dataset_for_brits import DatasetForBRITS
-from pypots.data.dataset_for_grud import DatasetForGRUD
-from pypots.data.dataset_for_mit import DatasetForMIT
 from pypots.data.generating import (
     gene_complete_random_walk,
     gene_random_walk_for_classification,
@@ -29,10 +26,6 @@ from pypots.data.utils import (
 __all__ = [
     # datasets
     "BaseDataset",
-    "DatasetForMIT",
-    "DatasetForBRITS",
-    "DatasetForGRUD",
-    "DatasetForGRUD",
     # data generation
     "gene_complete_random_walk",
     "gene_random_walk_for_classification",
