+++ pypots/forecasting/bttf/__init__.py
@@ -1,0 +1,12 @@
+"""
+
+"""
+
+# Created by Wenjie Du <wenjay.du@gmail.com>
+# License: GLP-v3
+
+from pypots.forecasting.bttf.model import BTTF
+
+__all__ = [
+    "BTTF",
+]
