+++ pypots/imputation/brits/__init__.py
@@ -1,0 +1,12 @@
+"""
+
+"""
+
+# Created by Wenjie Du <wenjay.du@gmail.com>
+# License: GLP-v3
+
+from pypots.imputation.brits.model import BRITS
+
+__all__ = [
+    "BRITS",
+]
