+++ pypots/imputation/locf/__init__.py
@@ -1,0 +1,12 @@
+"""
+
+"""
+
+# Created by Wenjie Du <wenjay.du@gmail.com>
+# License: GLP-v3
+
+from pypots.imputation.locf.model import LOCF
+
+__all__ = [
+    "LOCF",
+]
