+++ pypots/imputation/saits/model.py
@@ -1,5 +1,5 @@
 """
-PyTorch Transformer model for the time-series imputation task.
+PyTorch SAITS model for the time-series imputation task.
 
 Notes
 -----
@@ -19,183 +19,13 @@ import torch.nn.functional as F
 from torch.utils.data import DataLoader
 
 from pypots.data.base import BaseDataset
-from pypots.data.dataset_for_mit import DatasetForMIT
 from pypots.imputation.base import BaseNNImputer
+from pypots.imputation.saits.dataset import DatasetForSAITS
+from pypots.imputation.transformer.module import EncoderLayer, PositionalEncoding
 from pypots.utils.metrics import cal_mae
 
 
-class ScaledDotProductAttention(nn.Module):
-    """Scaled dot-product attention"""
-
-    def __init__(self, temperature: float, attn_dropout: float = 0.1):
-        super().__init__()
-        self.temperature = temperature
-        self.dropout = nn.Dropout(attn_dropout)
-
-    def forward(
-        self,
-        q: torch.Tensor,
-        k: torch.Tensor,
-        v: torch.Tensor,
-        attn_mask: torch.Tensor = None,
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        attn = torch.matmul(q / self.temperature, k.transpose(2, 3))
-        if attn_mask is not None:
-            attn = attn.masked_fill(attn_mask == 1, -1e9)
-        attn = self.dropout(F.softmax(attn, dim=-1))
-        output = torch.matmul(attn, v)
-        return output, attn
-
-
-class MultiHeadAttention(nn.Module):
-    """original Transformer multi-head attention"""
-
-    def __init__(
-        self,
-        n_head: int,
-        d_model: int,
-        d_k: int,
-        d_v: int,
-        attn_dropout: float,
-    ):
-        super().__init__()
-
-        self.n_head = n_head
-        self.d_k = d_k
-        self.d_v = d_v
-
-        self.w_qs = nn.Linear(d_model, n_head * d_k, bias=False)
-        self.w_ks = nn.Linear(d_model, n_head * d_k, bias=False)
-        self.w_vs = nn.Linear(d_model, n_head * d_v, bias=False)
-
-        self.attention = ScaledDotProductAttention(d_k**0.5, attn_dropout)
-        self.fc = nn.Linear(n_head * d_v, d_model, bias=False)
-
-    def forward(
-        self,
-        q: torch.Tensor,
-        k: torch.Tensor,
-        v: torch.Tensor,
-        attn_mask: torch.Tensor = None,
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        d_k, d_v, n_head = self.d_k, self.d_v, self.n_head
-        sz_b, len_q, len_k, len_v = q.size(0), q.size(1), k.size(1), v.size(1)
-
-        # Pass through the pre-attention projection: b x lq x (n*dv)
-        # Separate different heads: b x lq x n x dv
-        q = self.w_qs(q).view(sz_b, len_q, n_head, d_k)
-        k = self.w_ks(k).view(sz_b, len_k, n_head, d_k)
-        v = self.w_vs(v).view(sz_b, len_v, n_head, d_v)
-
-        # Transpose for attention dot product: b x n x lq x dv
-        q, k, v = q.transpose(1, 2), k.transpose(1, 2), v.transpose(1, 2)
-
-        if attn_mask is not None:
-            # this mask is imputation mask, which is not generated from each batch, so needs broadcasting on batch dim
-            attn_mask = attn_mask.unsqueeze(0).unsqueeze(
-                1
-            )  # For batch and head axis broadcasting.
-
-        v, attn_weights = self.attention(q, k, v, attn_mask)
-
-        # Transpose to move the head dimension back: b x lq x n x dv
-        # Combine the last two dimensions to concatenate all the heads together: b x lq x (n*dv)
-        v = v.transpose(1, 2).contiguous().view(sz_b, len_q, -1)
-        v = self.fc(v)
-        return v, attn_weights
-
-
-class PositionWiseFeedForward(nn.Module):
-    def __init__(self, d_in: int, d_hid: int, dropout: float = 0.1):
-        super().__init__()
-        self.w_1 = nn.Linear(d_in, d_hid)
-        self.w_2 = nn.Linear(d_hid, d_in)
-        self.layer_norm = nn.LayerNorm(d_in, eps=1e-6)
-        self.dropout = nn.Dropout(dropout)
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        residual = x
-        x = self.layer_norm(x)
-        x = self.w_2(F.relu(self.w_1(x)))
-        x = self.dropout(x)
-        x += residual
-        return x
-
-
-class EncoderLayer(nn.Module):
-    def __init__(
-        self,
-        d_time: int,
-        d_feature: int,
-        d_model: int,
-        d_inner: int,
-        n_head: int,
-        d_k: int,
-        d_v: int,
-        dropout: float = 0.1,
-        attn_dropout: float = 0.1,
-        diagonal_attention_mask: bool = False,
-    ):
-        super().__init__()
-
-        self.diagonal_attention_mask = diagonal_attention_mask
-        self.d_time = d_time
-        self.d_feature = d_feature
-
-        self.layer_norm = nn.LayerNorm(d_model)
-        self.slf_attn = MultiHeadAttention(n_head, d_model, d_k, d_v, attn_dropout)
-        self.dropout = nn.Dropout(dropout)
-        self.pos_ffn = PositionWiseFeedForward(d_model, d_inner, dropout)
-
-    def forward(self, enc_input: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
-        if self.diagonal_attention_mask:
-            mask_time = torch.eye(self.d_time).to(enc_input.device)
-        else:
-            mask_time = None
-
-        residual = enc_input
-        # here we apply LN before attention cal, namely Pre-LN, refer paper https://arxiv.org/abs/2002.04745
-        enc_input = self.layer_norm(enc_input)
-        enc_output, attn_weights = self.slf_attn(
-            enc_input, enc_input, enc_input, attn_mask=mask_time
-        )
-        enc_output = self.dropout(enc_output)
-        enc_output += residual
-
-        enc_output = self.pos_ffn(enc_output)
-        return enc_output, attn_weights
-
-
-class PositionalEncoding(nn.Module):
-    def __init__(self, d_hid: int, n_position: int = 200):
-        super().__init__()
-        # Not a parameter
-        self.register_buffer(
-            "pos_table", self._get_sinusoid_encoding_table(n_position, d_hid)
-        )
-
-    @staticmethod
-    def _get_sinusoid_encoding_table(n_position: int, d_hid: int) -> torch.Tensor:
-        """Sinusoid position encoding table"""
-
-        def get_position_angle_vec(position):
-            return [
-                position / np.power(10000, 2 * (hid_j // 2) / d_hid)
-                for hid_j in range(d_hid)
-            ]
-
-        sinusoid_table = np.array(
-            [get_position_angle_vec(pos_i) for pos_i in range(n_position)]
-        )
-        sinusoid_table[:, 0::2] = np.sin(sinusoid_table[:, 0::2])  # dim 2i
-        sinusoid_table[:, 1::2] = np.cos(sinusoid_table[:, 1::2])  # dim 2i+1
-        return torch.FloatTensor(sinusoid_table).unsqueeze(0)
-
-    def forward(self, x):
-        return x + self.pos_table[:, : x.size(1)].clone().detach()
-
-
-class _TransformerEncoder(nn.Module):
+class _SAITS(nn.Module):
     def __init__(
         self,
         n_layers: int,
@@ -207,6 +37,7 @@ class _TransformerEncoder(nn.Module):
         d_k: int,
         d_v: int,
         dropout: float,
+        diagonal_attention_mask: bool = True,
         ORT_weight: float = 1,
         MIT_weight: float = 1,
     ):
@@ -216,7 +47,7 @@ class _TransformerEncoder(nn.Module):
         self.ORT_weight = ORT_weight
         self.MIT_weight = MIT_weight
 
-        self.layer_stack = nn.ModuleList(
+        self.layer_stack_for_first_block = nn.ModuleList(
             [
                 EncoderLayer(
                     d_time,
@@ -228,31 +59,84 @@ class _TransformerEncoder(nn.Module):
                     d_v,
                     dropout,
                     0,
-                    False,
+                    diagonal_attention_mask,
+                )
+                for _ in range(n_layers)
+            ]
+        )
+        self.layer_stack_for_second_block = nn.ModuleList(
+            [
+                EncoderLayer(
+                    d_time,
+                    actual_d_feature,
+                    d_model,
+                    d_inner,
+                    n_head,
+                    d_k,
+                    d_v,
+                    dropout,
+                    0,
+                    diagonal_attention_mask,
                 )
                 for _ in range(n_layers)
             ]
         )
 
-        self.embedding = nn.Linear(actual_d_feature, d_model)
-        self.position_enc = PositionalEncoding(d_model, n_position=d_time)
         self.dropout = nn.Dropout(p=dropout)
-        self.reduce_dim = nn.Linear(d_model, d_feature)
+        self.position_enc = PositionalEncoding(d_model, n_position=d_time)
+        # for operation on time dim
+        self.embedding_1 = nn.Linear(actual_d_feature, d_model)
+        self.reduce_dim_z = nn.Linear(d_model, d_feature)
+        # for operation on measurement dim
+        self.embedding_2 = nn.Linear(actual_d_feature, d_model)
+        self.reduce_dim_beta = nn.Linear(d_model, d_feature)
+        self.reduce_dim_gamma = nn.Linear(d_feature, d_feature)
+        # for delta decay factor
+        self.weight_combine = nn.Linear(d_feature + d_time, d_feature)
 
-    def _process(self, inputs: dict) -> Tuple[torch.Tensor, torch.Tensor]:
+    def _process(self, inputs: dict) -> Tuple[torch.Tensor, list]:
         X, masks = inputs["X"], inputs["missing_mask"]
-        input_X = torch.cat([X, masks], dim=2)
-        input_X = self.embedding(input_X)
-        enc_output = self.dropout(self.position_enc(input_X))
-
-        for encoder_layer in self.layer_stack:
+        # first DMSA block
+        input_X_for_first = torch.cat([X, masks], dim=2)
+        input_X_for_first = self.embedding_1(input_X_for_first)
+        enc_output = self.dropout(
+            self.position_enc(input_X_for_first)
+        )  # namely, term e in the math equation
+        for encoder_layer in self.layer_stack_for_first_block:
             enc_output, _ = encoder_layer(enc_output)
 
-        learned_presentation = self.reduce_dim(enc_output)
-        imputed_data = (
-            masks * X + (1 - masks) * learned_presentation
-        )  # replace non-missing part with original data
-        return imputed_data, learned_presentation
+        X_tilde_1 = self.reduce_dim_z(enc_output)
+        X_prime = masks * X + (1 - masks) * X_tilde_1
+
+        # second DMSA block
+        input_X_for_second = torch.cat([X_prime, masks], dim=2)
+        input_X_for_second = self.embedding_2(input_X_for_second)
+        enc_output = self.position_enc(
+            input_X_for_second
+        )  # namely term alpha in math algo
+        for encoder_layer in self.layer_stack_for_second_block:
+            enc_output, attn_weights = encoder_layer(enc_output)
+
+        X_tilde_2 = self.reduce_dim_gamma(F.relu(self.reduce_dim_beta(enc_output)))
+
+        # attention-weighted combine
+        attn_weights = attn_weights.squeeze(dim=1)  # namely term A_hat in Eq.
+        if len(attn_weights.shape) == 4:
+            # if having more than 1 head, then average attention weights from all heads
+            attn_weights = torch.transpose(attn_weights, 1, 3)
+            attn_weights = attn_weights.mean(dim=3)
+            attn_weights = torch.transpose(attn_weights, 1, 2)
+
+        # namely term eta
+        combining_weights = torch.sigmoid(
+            self.weight_combine(torch.cat([masks, attn_weights], dim=2))
+        )
+        # combine X_tilde_1 and X_tilde_2
+        X_tilde_3 = (1 - combining_weights) * X_tilde_2 + combining_weights * X_tilde_1
+        # replace non-missing part with original data
+        X_c = masks * X + (1 - masks) * X_tilde_3
+
+        return X_c, [X_tilde_1, X_tilde_2, X_tilde_3]
 
     def impute(self, inputs: dict) -> torch.Tensor:
         imputed_data, _ = self._process(inputs)
@@ -260,12 +144,15 @@ class _TransformerEncoder(nn.Module):
 
     def forward(self, inputs: dict) -> dict:
         X, masks = inputs["X"], inputs["missing_mask"]
-        imputed_data, learned_presentation = self._process(inputs)
-        ORT_loss = cal_mae(learned_presentation, X, masks)
+        ORT_loss = 0
+        imputed_data, [X_tilde_1, X_tilde_2, X_tilde_3] = self._process(inputs)
 
-        MIT_loss = cal_mae(
-            learned_presentation, inputs["X_intact"], inputs["indicating_mask"]
-        )
+        ORT_loss += cal_mae(X_tilde_1, X, masks)
+        ORT_loss += cal_mae(X_tilde_2, X, masks)
+        ORT_loss += cal_mae(X_tilde_3, X, masks)
+        ORT_loss /= 3
+
+        MIT_loss = cal_mae(X_tilde_3, inputs["X_intact"], inputs["indicating_mask"])
 
         # `loss` is always the item for backward propagating to update the model
         loss = self.ORT_weight * ORT_loss + self.MIT_weight * MIT_loss
@@ -274,12 +161,49 @@ class _TransformerEncoder(nn.Module):
             "imputed_data": imputed_data,
             "ORT_loss": ORT_loss,
             "MIT_loss": MIT_loss,
-            "loss": loss,
+            "loss": loss,  # will be used for backward propagating to update the model
         }
         return results
 
 
-class Transformer(BaseNNImputer):
+class SAITS(BaseNNImputer):
+    """
+    Parameters
+    ----------
+    n_steps
+    n_features
+    n_layers
+    d_model
+    d_inner
+    n_head
+    d_k
+    d_v
+    dropout
+    diagonal_attention_mask
+    ORT_weight
+    MIT_weight
+    batch_size
+    epochs
+    patience : int, default = None,
+        Leaving it default as None will disable the early-stopping.
+
+    learning_rate
+    weight_decay
+    num_workers
+    device
+
+    saving_path : str, default = None,
+        The path for automatically saving model checkpoints and tensorboard files (i.e. loss values recorded during
+        training into a tensorboard file). Will not save if not given.
+
+    model_saving_strategy : str or None, None or "best" or "better" , default = "best",
+        The strategy to save model checkpoints. It has to be one of [None, "best", "better"].
+        No model will be saved when it is set as None.
+        The "best" strategy will only automatically save the best model after the training finished.
+        The "better" strategy will automatically save the model during training whenever the model performs
+        better than in previous epochs.
+    """
+
     def __init__(
         self,
         n_steps: int,
@@ -290,7 +214,8 @@ class Transformer(BaseNNImputer):
         n_head: int,
         d_k: int,
         d_v: int,
-        dropout: float,
+        dropout: int or float,
+        diagonal_attention_mask: bool = True,
         ORT_weight: int = 1,
         MIT_weight: int = 1,
         batch_size: int = 32,
@@ -325,10 +250,11 @@ class Transformer(BaseNNImputer):
         self.d_k = d_k
         self.d_v = d_v
         self.dropout = dropout
+        self.diagonal_attention_mask = diagonal_attention_mask
         self.ORT_weight = ORT_weight
         self.MIT_weight = MIT_weight
 
-        self.model = _TransformerEncoder(
+        self.model = _SAITS(
             self.n_layers,
             self.n_steps,
             self.n_features,
@@ -338,13 +264,14 @@ class Transformer(BaseNNImputer):
             self.d_k,
             self.d_v,
             self.dropout,
+            self.diagonal_attention_mask,
             self.ORT_weight,
             self.MIT_weight,
         )
         self.model = self.model.to(self.device)
         self._print_model_size()
 
-    def _assemble_input_for_training(self, data: dict) -> dict:
+    def _assemble_input_for_training(self, data: list) -> dict:
         """Assemble the given data into a dictionary for training input.
 
         Parameters
@@ -371,7 +298,7 @@ class Transformer(BaseNNImputer):
 
         return inputs
 
-    def _assemble_input_for_validating(self, data: list) -> dict:
+    def _assemble_input_for_validating(self, data) -> dict:
         """Assemble the given data into a dictionary for validating input.
 
         Notes
@@ -395,10 +322,9 @@ class Transformer(BaseNNImputer):
             "X": X,
             "missing_mask": missing_mask,
         }
-
         return inputs
 
-    def _assemble_input_for_testing(self, data: list) -> dict:
+    def _assemble_input_for_testing(self, data) -> dict:
         """Assemble the given data into a dictionary for testing input.
 
         Notes
@@ -448,7 +374,7 @@ class Transformer(BaseNNImputer):
 
         """
         # Step 1: wrap the input data with classes Dataset and DataLoader
-        training_set = DatasetForMIT(train_set, file_type=file_type)
+        training_set = DatasetForSAITS(train_set, file_type=file_type)
         training_loader = DataLoader(
             training_set,
             batch_size=self.batch_size,
@@ -487,7 +413,11 @@ class Transformer(BaseNNImputer):
         # Step 3: save the model if necessary
         self.auto_save_model_if_necessary(training_finished=True)
 
-    def impute(self, X: Union[dict, str], file_type: str = "h5py") -> np.ndarray:
+    def impute(
+        self,
+        X: Union[dict, str],
+        file_type="h5py",
+    ) -> np.ndarray:
         """Impute missing values in the given data with the trained model.
 
         Parameters
@@ -504,6 +434,7 @@ class Transformer(BaseNNImputer):
         array-like, shape [n_samples, sequence length (time steps), n_features],
             Imputed data.
         """
+        # Step 1: wrap the input data with classes Dataset and DataLoader
         self.model.eval()  # set the model as eval status to freeze it.
         test_set = BaseDataset(X, return_labels=False, file_type=file_type)
         test_loader = DataLoader(
@@ -514,11 +445,13 @@ class Transformer(BaseNNImputer):
         )
         imputation_collector = []
 
+        # Step 2: process the data with the model
         with torch.no_grad():
             for idx, data in enumerate(test_loader):
                 inputs = self._assemble_input_for_testing(data)
                 imputed_data = self.model.impute(inputs)
                 imputation_collector.append(imputed_data)
 
+        # Step 3: output collection and return
         imputation_collector = torch.cat(imputation_collector)
         return imputation_collector.cpu().detach().numpy()
