+++ pypots/imputation/transformer/module.py
@@ -1,0 +1,188 @@
+"""
+PyTorch Transformer model for the time-series imputation task.
+
+Notes
+-----
+Partial implementation uses code from https://github.com/WenjieDu/SAITS.
+"""
+
+# Created by Wenjie Du <wenjay.du@gmail.com>
+# License: GPL-v3
+
+from typing import Tuple
+
+import numpy as np
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+
+class ScaledDotProductAttention(nn.Module):
+    """Scaled dot-product attention"""
+
+    def __init__(self, temperature: float, attn_dropout: float = 0.1):
+        super().__init__()
+        self.temperature = temperature
+        self.dropout = nn.Dropout(attn_dropout)
+
+    def forward(
+        self,
+        q: torch.Tensor,
+        k: torch.Tensor,
+        v: torch.Tensor,
+        attn_mask: torch.Tensor = None,
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        attn = torch.matmul(q / self.temperature, k.transpose(2, 3))
+        if attn_mask is not None:
+            attn = attn.masked_fill(attn_mask == 1, -1e9)
+        attn = self.dropout(F.softmax(attn, dim=-1))
+        output = torch.matmul(attn, v)
+        return output, attn
+
+
+class MultiHeadAttention(nn.Module):
+    """original Transformer multi-head attention"""
+
+    def __init__(
+        self,
+        n_head: int,
+        d_model: int,
+        d_k: int,
+        d_v: int,
+        attn_dropout: float,
+    ):
+        super().__init__()
+
+        self.n_head = n_head
+        self.d_k = d_k
+        self.d_v = d_v
+
+        self.w_qs = nn.Linear(d_model, n_head * d_k, bias=False)
+        self.w_ks = nn.Linear(d_model, n_head * d_k, bias=False)
+        self.w_vs = nn.Linear(d_model, n_head * d_v, bias=False)
+
+        self.attention = ScaledDotProductAttention(d_k**0.5, attn_dropout)
+        self.fc = nn.Linear(n_head * d_v, d_model, bias=False)
+
+    def forward(
+        self,
+        q: torch.Tensor,
+        k: torch.Tensor,
+        v: torch.Tensor,
+        attn_mask: torch.Tensor = None,
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        d_k, d_v, n_head = self.d_k, self.d_v, self.n_head
+        sz_b, len_q, len_k, len_v = q.size(0), q.size(1), k.size(1), v.size(1)
+
+        # Pass through the pre-attention projection: b x lq x (n*dv)
+        # Separate different heads: b x lq x n x dv
+        q = self.w_qs(q).view(sz_b, len_q, n_head, d_k)
+        k = self.w_ks(k).view(sz_b, len_k, n_head, d_k)
+        v = self.w_vs(v).view(sz_b, len_v, n_head, d_v)
+
+        # Transpose for attention dot product: b x n x lq x dv
+        q, k, v = q.transpose(1, 2), k.transpose(1, 2), v.transpose(1, 2)
+
+        if attn_mask is not None:
+            # this mask is imputation mask, which is not generated from each batch, so needs broadcasting on batch dim
+            attn_mask = attn_mask.unsqueeze(0).unsqueeze(
+                1
+            )  # For batch and head axis broadcasting.
+
+        v, attn_weights = self.attention(q, k, v, attn_mask)
+
+        # Transpose to move the head dimension back: b x lq x n x dv
+        # Combine the last two dimensions to concatenate all the heads together: b x lq x (n*dv)
+        v = v.transpose(1, 2).contiguous().view(sz_b, len_q, -1)
+        v = self.fc(v)
+        return v, attn_weights
+
+
+class PositionWiseFeedForward(nn.Module):
+    def __init__(self, d_in: int, d_hid: int, dropout: float = 0.1):
+        super().__init__()
+        self.w_1 = nn.Linear(d_in, d_hid)
+        self.w_2 = nn.Linear(d_hid, d_in)
+        self.layer_norm = nn.LayerNorm(d_in, eps=1e-6)
+        self.dropout = nn.Dropout(dropout)
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        residual = x
+        x = self.layer_norm(x)
+        x = self.w_2(F.relu(self.w_1(x)))
+        x = self.dropout(x)
+        x += residual
+        return x
+
+
+class EncoderLayer(nn.Module):
+    def __init__(
+        self,
+        d_time: int,
+        d_feature: int,
+        d_model: int,
+        d_inner: int,
+        n_head: int,
+        d_k: int,
+        d_v: int,
+        dropout: float = 0.1,
+        attn_dropout: float = 0.1,
+        diagonal_attention_mask: bool = False,
+    ):
+        super().__init__()
+
+        self.diagonal_attention_mask = diagonal_attention_mask
+        self.d_time = d_time
+        self.d_feature = d_feature
+
+        self.layer_norm = nn.LayerNorm(d_model)
+        self.slf_attn = MultiHeadAttention(n_head, d_model, d_k, d_v, attn_dropout)
+        self.dropout = nn.Dropout(dropout)
+        self.pos_ffn = PositionWiseFeedForward(d_model, d_inner, dropout)
+
+    def forward(self, enc_input: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor]:
+        if self.diagonal_attention_mask:
+            mask_time = torch.eye(self.d_time).to(enc_input.device)
+        else:
+            mask_time = None
+
+        residual = enc_input
+        # here we apply LN before attention cal, namely Pre-LN, refer paper https://arxiv.org/abs/2002.04745
+        enc_input = self.layer_norm(enc_input)
+        enc_output, attn_weights = self.slf_attn(
+            enc_input, enc_input, enc_input, attn_mask=mask_time
+        )
+        enc_output = self.dropout(enc_output)
+        enc_output += residual
+
+        enc_output = self.pos_ffn(enc_output)
+        return enc_output, attn_weights
+
+
+class PositionalEncoding(nn.Module):
+    def __init__(self, d_hid: int, n_position: int = 200):
+        super().__init__()
+        # Not a parameter
+        self.register_buffer(
+            "pos_table", self._get_sinusoid_encoding_table(n_position, d_hid)
+        )
+
+    @staticmethod
+    def _get_sinusoid_encoding_table(n_position: int, d_hid: int) -> torch.Tensor:
+        """Sinusoid position encoding table"""
+
+        def get_position_angle_vec(position):
+            return [
+                position / np.power(10000, 2 * (hid_j // 2) / d_hid)
+                for hid_j in range(d_hid)
+            ]
+
+        sinusoid_table = np.array(
+            [get_position_angle_vec(pos_i) for pos_i in range(n_position)]
+        )
+        sinusoid_table[:, 0::2] = np.sin(sinusoid_table[:, 0::2])  # dim 2i
+        sinusoid_table[:, 1::2] = np.cos(sinusoid_table[:, 1::2])  # dim 2i+1
+        return torch.FloatTensor(sinusoid_table).unsqueeze(0)
+
+    def forward(self, x):
+        return x + self.pos_table[:, : x.size(1)].clone().detach()
