+++ pypots/utils/__init__.py
@@ -1,12 +1,15 @@
 """
-
+Utility functions and classes are stored in this package.
 """
 
 # Created by Wenjie Du <wenjay.du@gmail.com>
 # License: GPL-v3
 
+
 __all__ = [
+    # content files in this package
     "files",
     "logging",
     "metrics",
+    "random",
 ]
